+++ main/infrastructure/requests/saved_objects_get_index_pattern_request.go
@@ -21,6 +21,7 @@ func (r SavedObjectsGetIndexPatternRequest) Do(
 
 	method = "GET"
 
+	// TODO: this is a OpenSearch Dashboards API. Without it, it doesn't work
 	relative := "_dashboards/api/saved_objects/_find?type=index-pattern"
 	path.Grow(1 + len(relative))
 	path.WriteString("/")