+++ main/infrastructure/requests/saved_objects_post_index_pattern_request.go
@@ -23,6 +23,7 @@ func (r SavedObjectsPostIndexPatternRequest) Do(
 
 	method = "POST"
 
+	// TODO: this is a OpenSearch Dashboards API. Without it, it doesn't work
 	relative := "_dashboards/api/saved_objects/index-pattern"
 	path.Grow(1 + len(relative))
 	path.WriteString("/")