+++ CHANGELOG.md
@@ -9,7 +9,7 @@ When a new release is proposed:
 1. Create a new branch `bump/x.x.x` (this isn't a long-lived branch!!!);
 2. The Unreleased section on `CHANGELOG.md` gets a version number and date;
 3. Open a Pull Request with the bump version changes targeting the `main` branch;
-5. When the Pull Request is merged a new git tag must be created using [GitLab environment](https://gitlab.com/intruderlabs/toolbox/intrusearch/-/tags).
+4. When the Pull Request is merged a new git tag must be created using [GitLab environment](https://gitlab.com/intruderlabs/toolbox/intrusearch/-/tags).
 
 Releases to productive environments should run from a tagged version.
 Exceptions are acceptable depending on the circumstances (critical bug fixes that can be cherry-picked, etc.).
@@ -22,7 +22,7 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ### Changed
 
--
+- corrected the error deserialization when it's not found response
 
 ### Removed